ai-agents/seo-agent/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.8
openai>=1.12.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
python-dotenv>=1.0.0
requests>=2.31.0
# Google Search Console API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
